+++ phase4-profile-entsog/src/main/java/com/helger/phase4/profile/entsog/ENTSOGPMode.java
@@ -134,9 +134,13 @@ public final class ENTSOGPMode
     final ETriState eReceptionAwareness = ETriState.TRUE;
     final ETriState eRetry = ETriState.TRUE;
     final int nMaxRetries = 1;
-    final int nRetryIntervalMS = 10_000;
+    final long nRetryIntervalMS = 10_000;
     final ETriState eDuplicateDetection = ETriState.TRUE;
-    return new PModeReceptionAwareness (eReceptionAwareness, eRetry, nMaxRetries, nRetryIntervalMS, eDuplicateDetection);
+    return new PModeReceptionAwareness (eReceptionAwareness,
+                                        eRetry,
+                                        nMaxRetries,
+                                        nRetryIntervalMS,
+                                        eDuplicateDetection);
   }
 
   /**
@@ -162,8 +166,16 @@ public final class ENTSOGPMode
                                          @Nonnull final IPModeIDProvider aPModeIDProvider,
                                          final boolean bPersist)
   {
-    final PModeParty aInitiator = new PModeParty (ENTSOG_PARTY_ID_TYPE, sInitiatorID, CAS4.DEFAULT_INITIATOR_URL, null, null);
-    final PModeParty aResponder = new PModeParty (ENTSOG_PARTY_ID_TYPE, sResponderID, CAS4.DEFAULT_RESPONDER_URL, null, null);
+    final PModeParty aInitiator = new PModeParty (ENTSOG_PARTY_ID_TYPE,
+                                                  sInitiatorID,
+                                                  CAS4.DEFAULT_INITIATOR_URL,
+                                                  null,
+                                                  null);
+    final PModeParty aResponder = new PModeParty (ENTSOG_PARTY_ID_TYPE,
+                                                  sResponderID,
+                                                  CAS4.DEFAULT_RESPONDER_URL,
+                                                  null,
+                                                  null);
 
     final PMode aPMode = new PMode (aPModeIDProvider.getPModeID (sInitiatorID, sResponderID),
                                     aInitiator,
